+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 25. New submit code button
 26. Fetch test cases: pbinfo.ro support
 27. Fetch test cases: kilonova.ro support
+28. Fetch test cases: codeforces.com support
 
 ### Fixed/Changes: